Don't run the system on a separate thread
SDL_Renderer does not work when used on a separate thread. Instead
of running the system loop in a separate thread, make the thread
send SDL_USEREVENT to the main loop and run the system tick once
in response.

SDL_AddTimer could have been used to despatch events, but it has
only 10ms granularity and uses threads anyway, so it is better to
just manage it ourselves.