Refactor rendering for stretch/aspect/pixel
Changes the resize handling to not try to resize the window. Instead,
black borders are used. There are two bools, keep_aspect and keep_pixels.
keep_pixels causes the system pixels to always be an integer multiple of
device pixels. keep_aspect causes system pixels to always be square.
keep_pixels implies keep_aspect.

If neither is true the texture is stretched to the window size.