Refactor all system code into a separate file
This moves all code directly dealing with the blit simulation to
a separate file. Main no longer needs to include any blit code
except for input.hpp for the button names, and does not need to
include user code at all. This gives much clearer separation of
concerns.

All threading works as before, but now it is handled by a separate
class. Just call run() and stop() on that class.

The class provides functions to set input. They are protected by
a mutex so you can call them any time.

In order to make video recording work there is a function to get
the raw framebuffer. The video recorder should be rewritten to not
need this.

Note that this breaks the VC solution files but everything else
should work.