Make system size be hires
System width and height are now defined in the System class. The
values are now 320x240 because this reflects the native screen
resolution. These values end up getting used more in the code.

The stretching mode is now a bit different too. KeepPixels scales
to multiples of 320x240. KeepPixelsLores scales to multiples of
160x120. KeepPixels is the default as it works for both screen
modes, but it makes bigger borders when resizing.

Finally the minimum window size is now 320x240. There isn't much
point making it smaller than that.