Create the framebuffer textures as STREAMING
This is the texture type designed to be frequently updated. It can
be locked and directly updated without requiring an extra buffer.
We could update it with SDL_LockTexture for greater performance,
but SDL_UpdateTexture has a fast path for this and will gracefully
fall back to the slow method.