Refactor renderer recording
Instead of managing state inside the renderer, just implement a
read_pixels function. This has to create a texture every time it
is called, but it is not a big performance hit.

A nice feature of this is that it can be called any time to get
the current picture. This means it could be used for a screenshot
function too.