Attempt fix for RGB/RGBA support ofr #13
Input file format was always intended to be auto-detected since image formats typically contain this information.
The --format parameter was designed for output format, but the engine presently only supports paletted, compressed data due to memory constraints.

As of this update asset-builder will attempt to index the colours used in an RGB or RGBA input image into a palette, if the number of used colours exceeds 255 (8bpp) then it will fail.