Write LCD SBGR to fix colour order
During a respin we flipped the Red and Blue LCD traces for routing convinience, this changeset replaces a temporary hack used to correct for this.

The previous method swapped the colours during the copy from 32blit's buffer to the LTDC buffer, this only worked in lores mode since hires uses the LTDC buffer directly.

The SBGR bit is set on the ST7272A LCD over SPI to correct for this in hardware, avoiding the need for any colour order hacks in 32blit or the STM32 HAL.