Oversample and correct joystick input
This changeset corrects the joystick axes to match those in SDL: ie the positive axis directions should be Right and Down to match screen coordinates.

While it does not make any effort to provide calibration for variation in joystick parts (we will have to look into this as more devices are assembled) it does include a deadzone with clamping to -1 0 and +1

The STM32s ADC oversampling is used and the input LSB discarded to considerably reduce input noise. Since we're running the ADC in 16bit mode there is no perceptible loss of positional accuracy.