Add crude backlight brightness support
This changeset configures TIM15 for controlling the backlight.

The lynchpin for this was that TIM15 appears to call out to HAL_TIM_PWM_MspInit when initialised. This missing function meant the TIM15 clock was never enabled.

Presently the prescalar, period and pulse values are a little arbitrary but appear to produce a stable backlight (high PWM frequency) with very fine-grained brightness control.

The user can set the backlight in code with blit::backlight which takes a float from 0.0 (off) to 1.0 (on).

The backlight is configured as Open Drain, Pull Up so that it fails on if the PWM isn't running. It's also initialised with a pulse of 0 and inverted so a larger pulse value equals a dimmer backlight.