Move wsl-flash to tools and simplify
wsl-flash is purely for those who want to build to the STM32 in WSL on Windows 10 without having to jump through hoops to flash the resulting binary files.

I've eliminated the need for a separate .bat file since DfuSeCommand.exe can be called directly in WSL.

wsl-flash has been moved to tools, and now requires a relative path to the .bin file. It will build a .dfu file alongside it (overwriting if it exists) and flash it.