Fix config for DAC/DMA
This changeset sets up the DAC/DMA for 16kHz audio output. This was tested using an 8k sample buffer.

Keeping the sample buffer up-to-date relies on a half and full complete interrupt from the DMA transfer.

At half complete the first half of the buffer can be filled while the latter half is being DMA'd.

At full complete the latter half of the buffer can be filled while the first half is being DMA'd.

This allows 16kHz (and potentially beyond) audio streaming directly from the SD card.

In practise we will be outputting audio into the buffer from our own engine and likely need a much smaller buffer.