Add battery status and system menu
The system menu needs a lot of work, since it will likely provide considerably more functionality than what's currently implemented.

The BQ24295 power management chip is brought online and reports charge and adapter status. Right now it's running in unattended mode, but it's possible for it to run in host mode with control over charging options from the device.

The battery voltage reading needs additional smoothing, since it will fluctuate quite significantly and make for a very jittery battery bar. It also needs scaled to reflect the non-linear discharge rate of the LiPo.

DFU and Shipping menu options probably need to be in a more advanced submenu, but for development/testing purposes they are convinient here.