Fix timers and tweens
This set of changes shakes up the timer and tween functionality in the 32blit API.

Chiefly both now at least somewhat work as expected.

Huge thanks to FManga on Discord for figuring out where I was going wrong with the timer implementation and suggesting the `init()` pattern as a fix.

A usage example to follow!