Include the whole SDL HAL library
libBlitHalSDL.a contains SDL_main, which is required by libSDL2.a.
But libBlitHalSDL.a also depends on libSDL2.a, so we have a circular
dependency. Which ever one is listed first will have some required
symbols stripped by the linker. In order to avoid this, just use
--whole-archive when linking libBlitHalSDL.a, because we almost
certainly need all of it anyway.

This whole situation happens because it is rather unusual to have
main() inside a static library, rather than the source of the final
executable. But due to what we're doing, we don't have much choice.

Fixes #26