Add STM32 CubeMX IOC file
I've taken pains to get the configuration this IOC file as close to our codebase as possible so it should be possible to build a new 32Blit-compatible ground-up project using it as a starting point.

It's also useful for configuring or tweaking the main 32Blit codebase, since the individual files should now be relatively easy to manually merge across.