Avoid std::rand for #33
This changeset avoids std::rand as suggested in issue #33.

This allows std::rand to be used in user code and give predictable pseudo-random results between device and SDL testbed.

blit::random is now implemented in SDL using a 32bit mesenne twister which is, in turn, seeded by the system random_device on Linux and by milliseconds since the epoch on MinGW/Windows (since random_device produces an unseeded deterministic sequence in this case)

This means blit::random now produces (mostly) non-deterministic random numbers on both the STM32 and other targets.