Copy SDL2.dll for Windows builds
This rather long-winded approach includes a work-around for the fact that SDL2_PREFIX doesn't appear to normally be available to the .toolchain file.

It now checks SDL2_DIR, sets it to the README.md mentioned default if not set and imports sdl2-config.cmake to set these variables.

With SDL2_PREFIX it's now possible to copy SDL2.dll into the build output folder automagically.

Maybe there's a better way to do this?