Overhaul scrolly-tile
I've made scrolly-tile a playable game of rapid ascent- it's really difficult not to end up with Celeste.

I have overhauled the procedural level generation to use 5 winding passages which are turned on/off to widen/narrow the overall level.

Passages are connected to each other during a narrowing, so the player is never left in the lurch- though there are still bugs!

Water is continuously chasing the player, it's always on screen and always rising but climbing will keep it at bay.

Right now players can triple-jump which makes it all rather easy. Much further tweaking required.

Enjoy the rainbows.