Working ADC DMA
This change completes the switch over from polling ADC to an interrupt-driven DMA setup.

ADC polling was previously calling hangups presumably with some register or issue with SysTick. And this method is nicer.

ADC3 currently has something of a hack to align the channel reads to the 32 sample circular buffer.