Allow custom name for assets
May not yet fix the pack_sprites function TODO, but allows the assets to be given names independent of their filename

It does this by popping 3 args at a time from blit_assets- type, file and name

Type and file are processed as usual, but names are collated and passed to asset-packer via a new argument: --names

This is a comma-separated list of asset names that must match the order of files supplied.

You can now use:

```
blit_assets (rocks-and-diamonds
    MAP_TILED2BIN level.tmx asset_level
    SPRITE_PACKED sprites.png asset_sprites)
```

Include `assets.hpp`

And then access `asset_level.data` and `asset_sprites.data` in your code.