Fix battery voltage ADC reading
Battery voltage has been broken for a while, probably in the migration from normal to interrupt-driven ADC.

I've added GPIO config for the relevant ADC channels, albeit they seemed to work without it. It's best to be explicit.

I've switched from 32.5 cycle integration time to 810.5. With 256 sample oversampling this means each of the four ADC3 readings takes 0.43ms, but since it's asyncronous that's largely irrelevant.

It seems - at an educated guess - the longer integration time is required to successfully read a voltage through the 1MOhm resistor divider pair.