DMA2d interrupt control (#511)
* Update display.cpp

now the dma2d work is controlled by interrupt. So we don't need to wait it to finish. A lot of CPU time could be saved

* Update display.hpp

add function to support statement check

* Update 32blit.cpp

check the dma2d and framebuff statement before and after render /flip

* Update display.hpp

add miss function

* Update display.cpp

To make the code more clear, it would be better to seperator flags into independent bool var. We don't need to save this 8 bytes....

* Update display.cpp

fix typo