Only delete old firmware when not launching something
I was having an issue where running firmware-update-v0.1.8.blit, or `make firmware-update.flash` would jump my 32blit into a semi broken state.

I believe - when not launching `firmware-update.blit` from SD card - that the firmware updater was being deleted from flash before it had a chance to be run.

This patch changes that behaviour so old updates are only deleted when the reset_target is 0.