Fix emscripten build for find_package()
Emscripten disables various automatic search paths because it is a
cross-compiler. One of the search paths it disables refers to the
source directory we are building. This means find_package with a path
relative to the source (ie ".") will not work correctly.

We can work around this by manually setting 32BLIT_DIR on the cmake
command.

Full explanation here:

https://github.com/emscripten-core/emscripten/issues/10078#issuecomment-580260069

...and various duplicate bugs on the Emscripten repository.