Add --help and --info commands, tweak output
This change adds a --help command to document --connect, --listen, --position and --info,
and additionally an --info command to output game metadata.

Games now output the title and version when launched, with a "Powered by 32Blit" line below.

Some reshuffling was needed so that --help and --info commands could bail before an SDL Window was initialised.